+++ example/lib/inputs/go_dynamic_input.dart
@@ -1,0 +1,19 @@
+
+import 'package:flutter/material.dart';
+import 'package:go_form/go_form.dart';
+
+class GoDynamicInput extends FormFieldModelBase<String>{
+  final String label;
+  GoDynamicInput( {required super.name, super.validator,required this.label,});
+
+  @override
+  Widget build(BuildContext context, FieldController controller) {
+    return TextFormField(
+      decoration: InputDecoration(
+        border: const OutlineInputBorder(),
+        errorText: controller.error,
+      ),
+      controller: controller.textController,
+    );
+  }
+}