+++ lib/src/controllers/field_controller.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import '../domain/domain.dart';
 
 class FieldController<T> extends ChangeNotifier {
+  TextEditingController? _textEditingController;
   T? get value => _valueNotifier.value.value.value;
   final ValueNotifier<FormFieldData<T?>> _valueNotifier;
 
@@ -17,11 +18,31 @@ class FieldController<T> extends ChangeNotifier {
           ),
         );
 
+
+  TextEditingController? get textController {
+    if (T == String) {
+      _textEditingController ??= TextEditingController(
+        text: _valueNotifier.value.value.value as String?,
+      );
+      _textEditingController!.addListener(() {
+        if (_textEditingController!.text != value) {
+          onChange(_textEditingController!.text as T?);
+        }
+      });
+    }
+    return _textEditingController;
+  }
+
   void onChange(T? newValue) {
     _valueNotifier.value = FormFieldData<T?>(
       initialValue: newValue,
       validator: _valueNotifier.value.validator,
     );
+    if (T == String && _textEditingController != null) {
+      if (_textEditingController!.text != newValue) {
+        _textEditingController!.text = newValue as String? ?? '';
+      }
+    }
     notifyListeners();
   }
 
@@ -31,6 +52,19 @@ class FieldController<T> extends ChangeNotifier {
       validator: _valueNotifier.value.validator,
       error: newError,
     );
+    notifyListeners();
+  }
+  void setValue(dynamic newValue) {
+    _valueNotifier.value=FormFieldData<T?>(
+      initialValue: newValue,
+      validator: _valueNotifier.value.validator,
+      error: _valueNotifier.value.error,
+    );
+    if (T == String && _textEditingController != null) {
+      if (_textEditingController!.text != newValue) {
+        _textEditingController!.text = newValue as String? ?? '';
+      }
+    }
     notifyListeners();
   }