+++ lib/src/domain/from_field_data.dart
@@ -21,4 +21,19 @@ class FormFieldData<T> {
   @override
   int get hashCode => value.hashCode^error.hashCode;
 
+  /// Creates a copy of this [FormFieldData] with the ability to override some properties.
+  FormFieldData<T> copyWith({
+    T? initialValue,
+    String? error,
+    String? Function(dynamic value)? validator,
+    T? value,
+  }) {
+    return FormFieldData<T>(
+      initialValue: initialValue ?? this.value.value,
+      error: error ?? this.error,
+      validator: validator ?? this.validator,
+    );
+  }
+
+
 }